refactor(application): relocate component_builder and dependency_builder tests to their modules
Move 9 component_builder tests and 4 dependency_builder tests from mod.rs
into component_builder.rs and dependency_builder.rs respectively so that
each file's tests live alongside the code they exercise.

No tests added or removed; all 39 existing tests continue to pass.

Closes #532
